First requests sent
Register form functionaly implemented.
The form has all the inputs necessary. It checks that the user fills in al the fields, confirms their password and that the password is 8 characters or longer.

Used a function called getData to initially test requests.

Then on submit the form collects all the data from the inputs and sends it to the backend.